refactor: extract global error handler into dedicated middleware #621
